expense-manager-backend/requirements.txt
alembic==1.7.3
attrs==21.2.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
PyJWT==2.2.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
Werkzeug==2.0.1
